perf: defer comics burner graph
